+++ packages/api/app/distribution-routes.js
@@ -10,6 +10,8 @@ const {
 } = require('../endpoints/distribution');
 const { isAccessTokenExpired } = require('../lib/token');
 
+const version = require('../endpoints/version');
+
 /**
  * Helper function to pull bucket out of a path string.
  * Will ignore leading slash.
@@ -107,8 +109,6 @@ const s3Credentials = (req, res) => res.status(501).end();
 
 const s3CredentialsREADME = (req, res) => res.status(501).end();
 
-const version = (req, res) => res.status(501).end();
-
 router.get('/', root);
 router.get('/locate', locate);
 router.get('/login', login);
@@ -118,7 +118,8 @@ router.get('/pubkey', pubkey);
 router.get('/redirect', handleRedirectRequest);
 router.get('/s3Credentials', s3Credentials);
 router.get('/s3CredentialsREADME', s3CredentialsREADME);
-router.get('/version', version);
+// Use router.use to leverage custom version middleware
+router.use('/version', version);
 
 // HEAD /*
 // GET /* <- Actual presigned URL